ott-admin/backend/dsm_store/requirements.txt
fastapi==0.71.0
loguru==0.6.0
redis-py-cluster==2.1.3
uvicorn==0.16.0
pytest==7.0.1
boto3
kinesis-python
numpy
pandas